+++ circuits/src/field/native/native_chip.rs
@@ -1267,6 +1267,51 @@ where
         Ok(AssignedBit(res))
     }
 
+    fn is_not_equal(
+        &self,
+        layouter: &mut impl Layouter<F>,
+        x: &AssignedNative<F>,
+        y: &AssignedNative<F>,
+    ) -> Result<AssignedBit<F>, Error> {
+        // We enforce (i) (x - y) * aux = res
+        // and       (ii) (x - y) * (1 - res) = 0.
+        //  * If  x = y, (i)  implies res = 0; (ii) becomes trivial.
+        //  * If x != y, (ii) implies res = 1; (i) can be relaxed with aux = (x - y)^-1.
+
+        let value_cols = &self.config.value_cols;
+        let res_val = x.value().zip(y.value()).map(|(x, y)| F::from((x != y) as u64));
+        let aux_val = x.value().zip(y.value()).map(|(x, y)| (*x - *y).invert().unwrap_or(F::ONE));
+
+        // (i) enforced as aux * x - aux * y - res = 0.
+        let res = layouter.assign_region(
+            || "is_not_equal (i)",
+            |mut region| {
+                region.assign_advice(|| "aux", value_cols[0], 0, || aux_val)?;
+                self.copy_in_row(&mut region, x, &value_cols[1], 0)?;
+                self.copy_in_row(&mut region, y, &value_cols[2], 0)?;
+                let res = region.assign_advice(|| "res", value_cols[4], 0, || res_val)?;
+                let mut coeffs = [F::ZERO; NB_ARITH_COLS];
+                coeffs[4] = -F::ONE; // coeff of res
+                self.custom(&mut region, &coeffs, F::ZERO, (F::ONE, -F::ONE), F::ZERO, 0)?;
+                Ok(res)
+            },
+        )?;
+
+        // (ii) enforced as x - y - res * x + res * y = 0.
+        let must_be_zero = self.add_and_double_mul(
+            layouter,
+            (F::ZERO, &res),
+            (F::ONE, x),
+            (-F::ONE, y),
+            F::ZERO,
+            (-F::ONE, F::ONE),
+        )?;
+        self.assert_zero(layouter, &must_be_zero)?;
+
+        // The two equations we have enforced guarantee the bit-ness of `res`.
+        Ok(AssignedBit(res))
+    }
+
     fn is_equal_to_fixed(
         &self,
         layouter: &mut impl Layouter<F>,
@@ -1311,6 +1356,45 @@ where
         // The two equations we have enforced guarantee the bit-ness of `res`.
         Ok(AssignedBit(res))
     }
+
+    fn is_not_equal_to_fixed(
+        &self,
+        layouter: &mut impl Layouter<F>,
+        x: &AssignedNative<F>,
+        c: F,
+    ) -> Result<AssignedBit<F>, Error> {
+        // We enforce (i) (x - c) * aux = res
+        // and       (ii) (x - c) * (1 - res) = 0.
+        //  * If  x = c, (i)  implies res = 0; (ii) becomes trivial.
+        //  * If x != c, (ii) implies res = 1; (i) can be relaxed with aux = (x - c)^-1.
+
+        let value_cols = &self.config.value_cols;
+        let res_val = x.value().map(|x| F::from((*x != c) as u64));
+        let aux_val = x.value().map(|x| (*x - c).invert().unwrap_or(F::ONE));
+
+        // (i) enforced as - c * aux + aux * x - res = 0.
+        let res = layouter.assign_region(
+            || "is_not_equal (i)",
+            |mut region| {
+                region.assign_advice(|| "aux", value_cols[0], 0, || aux_val)?;
+                self.copy_in_row(&mut region, x, &value_cols[1], 0)?;
+                let res = region.assign_advice(|| "res", value_cols[4], 0, || res_val)?;
+                let mut coeffs = [F::ZERO; NB_ARITH_COLS];
+                coeffs[0] = -c; // coeff of aux
+                coeffs[4] = -F::ONE; // coeff of res
+                self.custom(&mut region, &coeffs, F::ZERO, (F::ONE, F::ZERO), F::ZERO, 0)?;
+                Ok(res)
+            },
+        )?;
+
+        // (ii) enforced as x - c + c * res - res * x = 0.
+        let must_be_zero =
+            self.add_and_mul(layouter, (c, &res), (F::ONE, x), (F::ZERO, x), -c, -F::ONE)?;
+        self.assert_zero(layouter, &must_be_zero)?;
+
+        // The two equations we have enforced guarantee the bit-ness of `res`.
+        Ok(AssignedBit(res))
+    }
 }
 
 impl<F> EqualityInstructions<F, AssignedBit<F>> for NativeChip<F>
@@ -1335,6 +1419,24 @@ where
         .map(AssignedBit)
     }
 
+    fn is_not_equal(
+        &self,
+        layouter: &mut impl Layouter<F>,
+        a: &AssignedBit<F>,
+        b: &AssignedBit<F>,
+    ) -> Result<AssignedBit<F>, Error> {
+        // a + b - 2ab
+        self.add_and_mul(
+            layouter,
+            (F::ONE, &a.0),
+            (F::ONE, &b.0),
+            (F::ZERO, &a.0),
+            F::ZERO,
+            -F::from(2),
+        )
+        .map(AssignedBit)
+    }
+
     fn is_equal_to_fixed(
         &self,
         layouter: &mut impl Layouter<F>,
@@ -1343,6 +1445,16 @@ where
     ) -> Result<AssignedBit<F>, Error> {
         let assigned_constant = self.assign_fixed(layouter, constant)?;
         self.is_equal(layouter, b, &assigned_constant)
+    }
+
+    fn is_not_equal_to_fixed(
+        &self,
+        layouter: &mut impl Layouter<F>,
+        b: &AssignedBit<F>,
+        constant: bool,
+    ) -> Result<AssignedBit<F>, Error> {
+        let assigned_constant = self.assign_fixed(layouter, constant)?;
+        self.is_not_equal(layouter, b, &assigned_constant)
     }
 }